Module Async.Interval

Parameters

I : INTERVAL

Signature

val forever : ms:int -> int t

Creates an infinite async stream that enumerates the natural numbers every ms.

val make : ms:int -> int connection

Creates an async stream that enumerates the natural numbers every ms. This is a connection-based version that get closed (no more signals are sent) when the output stream is called.