Lambda_streams

Lambda_streams is a simple streaming library based on using lambdas and closures as a basis to write simple, composable streams. It's inspired by callbag.

Generic streams

Sync

Lambda_streams.Sync streams

Async

Lambda_streams.Async streams

Finite streams

Finite.Sync

Lambda_streams.Finite.Sync streams

Finite.Async

Lambda_streams.Finite.Async streams